rename ProfilePage placeholder file to avoid potential conflicts
