+++ frontend/src/router/Router.tsx
@@ -1,7 +1,7 @@
 import { RouterProvider, createBrowserRouter } from 'react-router-dom';
 import App from '../App';
 import NotFound from '../pages/NotFound';
-import ProfilePage from '../pages/profile/ProfilePage';
+import ProfilePage from '../pages/profile/ProfilePagePlaceholder';
 import ProfilePublicInfoPage from '../pages/profile/ProfilePublicInfoPage';
 
 const routes = createBrowserRouter([
